+++ Packages/com.mimylab.fukuroudon/Runtime/ManualObjectSync/Scripts/MOSSceneObserver.cs
@@ -12,7 +12,7 @@ namespace MimyLab.FukuroUdon
 #if UNITY_EDITOR
     using UnityEditor;
     using UnityEditor.SceneManagement;
-    using UdonSharpEditor;
+    //using UdonSharpEditor;
 #endif
 
     [ExecuteAlways]