+++ Packages/com.mimylab.fukuroudon/Runtime/ManualObjectSync/Scripts/MOSUpdateManager.cs
@@ -14,7 +14,7 @@ namespace MimyLab.FukuroUdon
 #if UNITY_EDITOR
     using UnityEditor;
     using UnityEditor.SceneManagement;
-    using UdonSharpEditor;
+    //using UdonSharpEditor;
 #endif
 
     [Icon(ComponentIconPath.FukuroUdon)]