+++ Packages/com.mimylab.fukuroudon/Runtime/ManualObjectSync/Scripts/PickupEventTransfer.cs
@@ -7,10 +7,10 @@ https://opensource.org/licenses/mit-license.php
 namespace MimyLab.FukuroUdon
 {
     using UdonSharp;
-     using UnityEngine;
-     //using VRC.SDKBase;
-     using VRC.Udon;
-     //using VRC.SDK3.Components;
+    using UnityEngine;
+    using VRC.Udon;
+    //using VRC.SDK3.Components;
+
 
     [Icon(ComponentIconPath.FukuroUdon)]
     [AddComponentMenu("Fukuro Udon/Manual ObjectSync/Pickup Event Transfer")]