+++ Packages/com.mimylab.fukuroudon/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 利用手順は[こちら](https://github.com/mimyquality/FukuroUdon/wiki)からご確認ください。
 
+## [3.4.0-beta.3]
+
+- **Added**
+  - Manual ObjectSync に 新 SendCustomNetworkEvent 対応メソッドを追加しました。
+  - GameObject Celler に 新 SendCustomNetworkEvent 対応メソッドを追加しました。
+
 ## [3.4.0-beta.2]
 
 - **Added**
@@ -634,6 +640,7 @@
     - 0.2.7バージョンで、later-joiner視点でjoin直後が非アクティブだと同期しなかったのを修正しました。
     - 他人がピックアップした際にオブジェクトが一瞬跳ねる現象を低減しました。
 
+[3.4.0-beta.3]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.4.0-beta.3
 [3.4.0-beta.2]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.4.0-beta.2
 [3.4.0-beta.1]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.4.0-beta.1
 [3.3.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.3.0